Refactored some code, create separate file for actions on the post
