+++ app/src/main/java/com/example/twitterclonegpt/ui/homescreen/HomeScreen.kt
@@ -3,20 +3,11 @@ package com.example.twitterclonegpt.ui.homescreen
 import android.annotation.SuppressLint
 import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
-import dagger.hilt.android.AndroidEntryPoint
 
 @SuppressLint("UnusedMaterialScaffoldPaddingParameter")
 @Composable
 fun HomeScreen() {
     val viewModel: HomeScreenViewModel = hiltViewModel()
     Scaffold { HomeScreenContent(viewModel) }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun HomeScreenPreview() {
-//    HomeScreen()
 }
